컨트롤러에 post 요청 추가 cors 이슈 해결

+++ src/main/java/com/gamecrew/demo/dto/PlayerDto.java
@@ -16,7 +16,7 @@ public class PlayerDto {
 
     private String tag;
     private String name;
-    private Long trophies;
+    private int trophies;
 
     private List<BrawlersDto> brawlers;
 }
